+++ public/js/utils.js
@@ -1,4 +1,8 @@
 const deleteElection = async (electionId, csrf) => {
+    const conf = confirm("Are you sure you want to delete this election?");
+    if (!conf) {
+        return;
+    }
     fetch(`/elections/${electionId}`, {
         method: "DELETE",
         headers: {
@@ -12,4 +16,23 @@ const deleteElection = async (electionId, csrf) => {
             location.reload();
         }
     });
+};
+const hideMessage = () => {
+    const message = document.getElementById("message");
+    message.classList.remove("bottom-8");
+    message.classList.add("-bottom-full");
+};
+const handleMessages = (messages) => {
+    const message = document.getElementById("message");
+    message.classList.add("bottom-8");
+    message.classList.remove("-bottom-full");
+    const messageList = document.getElementById("message-list");
+    messageList.innerHTML = "";
+    console.log(messages);
+    messages.errors.forEach((message) => {
+        const li = document.createElement("li");
+        li.innerText = message.message;
+        messageList.appendChild(li);
+    });
+    setTimeout(hideMessage, 5000);
 };
